moved TestType enum to 'enums' folder
