+++ src/components/TestResults/HighScores.tsx
@@ -1,14 +1,13 @@
 import React, { useEffect, useState } from "react";
-import { TestType } from "../../App";
 import { getHighScores } from "../../firebase/GET/scoreGets";
 import { TimedScoreDocument, WordCountScoreDocument } from "../../firebase/firestoreDocumentInterfaces";
+import { TestType } from "../../enums";
 
 export interface HighScoresProps {
     isTestSubmitted: boolean,
     testType: TestType,
     testLength: number
 }
-
 
 const HighScores = ({isTestSubmitted, testType, testLength}: HighScoresProps) => {
 	const [highScoresArray, setHighScoresArray] = useState<TimedScoreDocument[] | WordCountScoreDocument[]>([]);
