+++ src/enums.ts
@@ -1,0 +1,4 @@
+export enum TestType {
+	Words = "words",
+	Time = "time"
+}
